requirements.txt
pandas
numpy
ta
python-telegram-bot
python-dotenv
yfinance
psycopg2-binary
streamlit
pandas
numpy
python-dotenv
matplotlib
python-dateutil
pytz
